Fixed some warnings from Codacy

+++ bignbr.c
@@ -876,11 +876,13 @@ void multint(BigInteger *pResult, const BigInteger *pMult, int factor)
     // In that case, there would be an error of +/- 1.
     if (low < HALF_INT_RANGE)
     {
-      carry = (int)(((double)(pLimb->x) * dFactor + (double)carry + HALF_INT_RANGE / 2)*dVal);
+      carry = (int)(((double)(pLimb->x) * dFactor + (double)carry +
+        (double)(HALF_INT_RANGE / 2))*dVal);
     }
     else
     {
-      carry = (int)(((double)(pLimb->x) * dFactor + (double)carry - HALF_INT_RANGE / 2)*dVal);
+      carry = (int)(((double)(pLimb->x) * dFactor + (double)carry -
+        (double)(HALF_INT_RANGE / 2))*dVal);
     }
     pResultLimb->x = low;
     pResultLimb++;
